wip: update cd_player example
