+++ cpp/examples/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(canonical)
 add_subdirectory(cd_player)
+add_subdirectory(cd_player_orig)
 add_subdirectory(controller)
