+++ cpp/examples/cd_player/CMakeLists.txt
@@ -1,6 +1,13 @@
-add_executable(cd_player
+add_library(cd_player_lib
         CDPlayer.cpp
 )
-target_link_libraries(cd_player
+target_link_libraries(cd_player_lib
         eta_hsm_lib
+)
+
+add_executable(cd_player
+        main.cpp
+)
+target_link_libraries(cd_player
+        cd_player_lib
 )
