+++ cpp/examples/cd_player_orig/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_executable(cd_player_orig
+        CDPlayer.cpp
+)
+target_link_libraries(cd_player_orig
+        eta_hsm_lib
+)
